move rr and subs files into /lib
