+++ lib/remoteResource.js
@@ -1,6 +1,6 @@
 const Mustache = require('mustache');
 
-const log = require('../lib/bunyan-api').createLogger('cluster-subscription');
+const log = require('./bunyan-api').createLogger('cluster-subscription');
 
 const { KubeClass, KubeApiConfig } = require('@razee/kubernetes-util');
 const kubeApiConfig = KubeApiConfig();
