+++ lib/subscriptions.js
@@ -1,4 +1,4 @@
-const log = require('../lib/bunyan-api').createLogger('cluster-subscription');
+const log = require('./bunyan-api').createLogger('cluster-subscription');
 const { execute } = require('apollo-link');
 const { WebSocketLink } = require('apollo-link-ws');
 const { SubscriptionClient } = require('subscriptions-transport-ws');
